server/requirements.txt
# Core Framework
fastapi==0.110.2
uvicorn==0.29.0
pydantic[email]==2.7.1

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Supabase
supabase==2.7.4

# Data Processing
pandas==2.2.2
numpy==1.26.4

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.21.0
scrapy==2.11.1
webdriver-manager==4.0.1

# Data Visualization
matplotlib==3.8.4
seaborn==0.13.2

# Templates & Configuration
jinja2==3.1.3
PyYAML==6.0.1

# Testing
pytest==8.2.2

# Development (Auto-reload)
watchdog[watchmedo]==3.0.0
pytest-watch==4.2.0

# Utilities
python-dotenv
